OK, look.
CreateJamView is subscribed to CreateJam changes. When you click "Make
this my jam", a new tab is created and switched to, which means that
CreateJam (which is subscribed to tab changes) updates its title and url
attributes. This triggers a re-render on CreateJamView and causes the
popup to stick around after you've clicked the button. Not good.

It took ages to find this bug because it seems dependent on a race
condition. Specifically, having tracked it down to the commit that
introduced it, which was a commit that only introduced a bit of markup
and CSS to the sign in screen (c715e24), and I was able to flip the bug
on and off by removing a single line of CSS ("margin-top: 16px").

I don't know either.